# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/tasks/create": {
            "post": {
                "description": "Create a new task with title and content",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Create a new task",
                "parameters": [
                    {
                        "description": "Create task request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_interfaces_task_model.CreateTaskReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task created successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response"
                        }
                    }
                }
            }
        },
        "/tasks/list": {
            "get": {
                "description": "Get all tasks for current user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Get task list",
                "responses": {
                    "200": {
                        "description": "Task list retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response"
                        }
                    }
                }
            }
        },
        "/tasks/recycle-list": {
            "get": {
                "description": "Get all deleted tasks in recycle bin",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Get recycle bin task list",
                "responses": {
                    "200": {
                        "description": "Recycle bin task list retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response"
                        }
                    }
                }
            }
        },
        "/tasks/update/{id}": {
            "put": {
                "description": "Update task title and content by task ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Update task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update task request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_interfaces_task_model.UpdateTaskReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task updated successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response"
                        }
                    }
                }
            }
        },
        "/tasks/update/{id}/status": {
            "put": {
                "description": "Update task status by task ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Update task status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Task status",
                        "name": "status",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task status updated successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response"
                        }
                    }
                }
            }
        },
        "/user/avatar": {
            "post": {
                "description": "Upload and update user avatar image",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update user avatar",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Avatar image file",
                        "name": "avatar",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Avatar updated successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "User login authentication",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "Login request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_interfaces_user_model.UserLoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_interfaces_user_model.UserInfoResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response"
                        }
                    }
                }
            }
        },
        "/user/logout": {
            "get": {
                "description": "User logout",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User logout",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response"
                        }
                    }
                }
            }
        },
        "/user/profile": {
            "get": {
                "description": "Get current user profile information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user information",
                "responses": {
                    "200": {
                        "description": "User information retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_interfaces_user_model.UserInfoResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response"
                        }
                    }
                }
            }
        },
        "/user/refresh-token": {
            "post": {
                "description": "Refresh User AccessToken",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Refresh User Token",
                "responses": {
                    "200": {
                        "description": "Refresh token successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "Register a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User registration",
                "parameters": [
                    {
                        "description": "Registration request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_interfaces_user_model.UserRegisterReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registration successful",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_interfaces_user_model.UserInfoResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response"
                        }
                    }
                }
            }
        },
        "/user/reset-password": {
            "post": {
                "description": "Reset user password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Reset user password",
                "parameters": [
                    {
                        "description": "Reset password request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_interfaces_user_model.UserResetPasswordReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Password reset successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_crazyfrankie_zrpc-todolist_interfaces_task_model.CreateTaskReq": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "github_com_crazyfrankie_zrpc-todolist_interfaces_task_model.UpdateTaskReq": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "github_com_crazyfrankie_zrpc-todolist_interfaces_user_model.UserInfoResp": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "user_create_time": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "github_com_crazyfrankie_zrpc-todolist_interfaces_user_model.UserLoginReq": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "github_com_crazyfrankie_zrpc-todolist_interfaces_user_model.UserRegisterReq": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "github_com_crazyfrankie_zrpc-todolist_interfaces_user_model.UserResetPasswordReq": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "github_com_crazyfrankie_zrpc-todolist_pkg_gin_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}